requirements.txt
# FastAgent core dependencies
litellm>=1.70.0
python-dotenv>=1.0.0
openai>=1.0.0
jsonschema>=4.25.0
mcp>=1.0.0
anthropic>=0.71.0
pillow>=12.0.0
pyyaml>=6.0.0
colorama>=0.4.6
websockets>=13.0.0

# Local server dependencies (cross-platform)
flask>=3.1.0
pyautogui>=0.9.54
pydantic>=2.12.0
requests>=2.32.0

# # macOS-specific dependencies (local server)
# pyobjc-core>=12.0; sys_platform == 'darwin'
# pyobjc-framework-cocoa>=12.0; sys_platform == 'darwin'
# pyobjc-framework-quartz>=12.0; sys_platform == 'darwin'
# atomacos>=3.2.0; sys_platform == 'darwin'

# # Linux-specific dependencies (local server)
# python-xlib>=0.33; sys_platform == 'linux'
# pyatspi>=2.38.0; sys_platform == 'linux'
# numpy>=1.24.0; sys_platform == 'linux'

# # Windows-specific dependencies (local server)
# pywinauto>=0.6.8; sys_platform == 'win32'
# pywin32>=306; sys_platform == 'win32'
# PyGetWindow>=0.0.9; sys_platform == 'win32'